+++ ptai-ee-tools-java/ptai-cli-plugin/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/cli/operations/LocalAstOperations.java
@@ -45,7 +45,7 @@ public class LocalAstOperations implements AstOperations {
     }
 
     public void scanStartedCallback(@NonNull final Project project, @NonNull UUID scanResultId) {
-        GracefulShutdown shutdown = new GracefulShutdown(owner);
+        shutdown = new GracefulShutdown(owner);
         Runtime.getRuntime().addShutdownHook(shutdown);
     }